Fix TestLocalStateGzip
Now that we run `ExecuteRegisterResourceOutputs` for the stack at the end of the deployment, it changes the behavior of snapshot writing, which requires an update to this test.

Before, we'd mutate the state for the stack's outputs in the middle of the deployment, causing the state to be saved in the middle of the deployment and then later when the deployment was complete, the stack would be saved again due to `hasElidedWrites` being true.

With the change to run `ExecuteRegisterResourceOutputs` at the end of the deployment, we're only saving the state once for this stack. All prior save attempts had `hasElidedWrites` being true, and then a final single save for the stack outputs, meaning there's no need for a second save.

Because of this, we now don't have `.json.gz.bak` file to check in this test.

To fix, we run `pulumi up` twice, to ensure the `.json.gz.bak` file is available to assert.