automation: only read complete lines before trying to deserialize
When tailing the event log in automation API we currently have nothing
that makes sure we read only complete lines.  This means if the OS
happens to flush an incomplete line for whatever reason (or the Go
JSON encoder does, which we're using to write these lines), we might
read a line that is incompletely written, and thus will fail to JSON
decode it.

Since the JSON encoder always writes a newline at the end of each
string, we can also make sure that the line we read ends with a
newline and otherwise wait for the rest of the line to be written.

The library we use in Go provides a convenient setting for this, while
in python and nodejs we need to add some code to do this ourselves.