revert readlines fix for nodejs; add test
I've been trying to add a test for this, but it looks like the library
actually deals with splitting the lines correctly.  Given that I don't
want to accidentailly hide any potential issues and bring the logging
back.

The test can still be useful for the future, so I left it there.