revert readlines fix for nodejs (#15798)
I've been trying to add a test for this, but it looks like the library
actually deals with splitting the lines correctly. Also reading back
through the issue again, it seems like the lines are missing the start
of the line, not the end, so this fix wouldn't work. Given that I don't
want to accidentally hide any potential issues and bring the logging
back.

The test can still be useful for the future, so I left it there.

I'm reopening https://github.com/pulumi/pulumi/issues/6768 as well.