fix checking for changes in on push job
In the on-push job, we are already on `master`.  Therefore diff
`master...HEAD` will never result in any changes.  Also we need to
deepen the fetch here so we have an actual previous commit to compare
with.

In addition to that, it looks like this actually results in a string
output not a boolean, so we need to check against that, or the
condition will always evaluate to true.