Run conformance tests with and without ts-node (#15809)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

Fixes https://github.com/pulumi/pulumi/issues/15075.

This doubles up the conformance tests to run with `tsc` instead of
`ts-node`. Getting this to work in the conformance framework required a
bit of pushing the envelope for what it means to "install dependencies"
and "run", but it works for testing, and isn't horribly exposed to
users.

It also points out a couple of things that we might want to clean up for
users wanting the same workflow.
1. Adding a `Build` command (which we might want for the type checker
work added in https://github.com/pulumi/pulumi/pull/15725 as well).
2. If you set "main" to ./bin because you're using tsc, you can't then
use InstallDependencies because to start with the "bin" folder is
missing.


## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [ ] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->