avoid racyness from the polling library (#15856)
The library we're using unfortunately has a race condition that's
especially worse when using polling to check for new changes of the
file. Avoid this racyness by sleeping for a bit for now, but really this
should be fixed upstream ideally.

I opened a upstream issue for this in
https://github.com/nxadm/tail/issues/67. I think we should merge this in
the meantime, as it makes this work, even though it's very hacky.

I tested this in https://github.com/pulumi/pulumi/pull/15853, where I
could no longer reproduce the failure we've seen before.

Fixes https://github.com/pulumi/pulumi/issues/15659
Fixes https://github.com/pulumi/pulumi/issues/15235