Allow Python dynamic providers to capture secrets
This allows the serialization used by dynamic providers to capture secrets, and if so, marks the resulting input value as a secret so that the `__provider` property will be a secret in the state.

The `__provider` property will no longer always be a secret. It will only be marked as a secret if a secret was captured during serialization of the provider.

Note: Previously, it wasn't possible to capture Outputs when serializing the resource provider. This change makes it possible and allows secrets to be captured.