Run integration tests and dev builds with race detection
Combining #15120 and #15124 and including fixes to allow tests to pass.

We use dev channel releases to catch bugs early. To increase the chance bugs are found, also enable race detection in these releases. This does mean these releases will be a little slower, and it might introduce more errors, but that seems like a good thing in exchange for potentially better quality of actual releases.

Enable race detection in the binary we're using for integration tests. This will allow us to catch more data races before they get into any release. This does mean the binary we're using for integration tests is slightly different from the binary we're releasing, however that's already the case as we're running a binary with coverage enabled for them. Later we rebuild the binary we're actually releasing.

This requires us to fix the race between snapshot code and the engine in RegisterResourceOutputs. I've done that by adding a lock to the State struct. This does not feel great, but it's a quick way to fix this and get race detection running (and unblocks #15871 which was also hitting the race detector because it started pulling snapshot code into unit tests as well).
There's probably a more principled overhaul that doesn't require locking at this level.

Fixes #15117