use test environments for codegen tests
Currently these tests always just use the directory in the repository
tree to run.  This isn't great for a couple of reasons:
- The tests write files into the repo tree (these are .gitignored, but
  still large and aren't cleaned up after the test finished)
- Tests running in parallel can potentially use the same directory.

We have a nice helper that creates new test environments and cleans
them up at the end of the tests.  Use that here for more test
isolation.

Also remove the t.Run() in the inner loop.  For a lot of these tests
it's not possible to run only one bit at a time (which is also why
there is an ordering enforced).  It's better to just make them run in
the outer test loop to avoid confusion when trying to rerun only part
of the test.