Lift context parameter to SerializeDeployment/Resource/Operations/Properties
SerializePropertyValue needed a `context.Context` object to pass to the
`config.Encrypter`. It was using `context.TODO()`, this change instead accepts
a context on the parameters and lifts that up to SerializeProperties,
SerializeResource, SerializeOperation, and SerializeDeployment.

There were a few call sites for those methods that already had a context on
hand, and they now pass that context. The other calls sites now use
`context.TODO()`, we should continue to iterate in this area to ensure
everywhere that needs a context has one passed in.