allow CopyFile to copy symlinks when we're running in Bazel
Bazel builds create symlinks of each file.  We still want to copy
these into test directories etc., so allow CopyFile to do that.  This
is a prerequisite for getting pulumi-yaml to build in Bazel.

We hide this with an environment variable, as that's easy to set
during Bazel builds, but does not affect this function during regular
usage.