Fix race condition when listing stacks (#15931)
When listing stacks, we currently try to get all the stacks from the
backend, and then get the checkpoint to for each stack to show the
summary.

However there's a race between listing the stacks and getting the
checkpoints, so the checkpoint may be missing by the time we get to that
part of the code.

Unfortunately the errors we're getting form getCheckpoint are not
structured, so we can't tell if we got a 404 from the cloud here, so the
next best option we have is re-checking whether the stack still exists.

This is another attempt at fixing
https://github.com/pulumi/pulumi/issues/15658.