Add missing `typescript` dependency to sxs_ts_test
With the release of v3.113.0 of `@pulumi/pulumi` we no longer have a dependency on `typescript`, so if someone has a dependency on `tsc`, they must specify it explicitly as a dependency. We're calling `tsc` for this test, so need to add an explicit dependency.

Note: This could be a `devDependency`, but the `package.json` files for the other test cases list the `typescript` dependency as a regular `dependency`, so doing the same here.