update dotnet SDK version in codegen tests (#15963)
Since we released a new SDK we should also update the version here.
Otherwise tests might end up failing as soon as a provider we're using
in the tests depends on the new dotnet SDK version.

Saw https://github.com/pulumi/pulumi/pull/15955 and realized we ran into
issues with this last time.